requirements.txt
ua_parser
asyncpg
gunicorn
uvloop; platform_system != "Windows"
pyyaml
aiohttp
aiohttp-apispec
marshmallow
sentry_sdk
passlib
psutil
cryptography